requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-together==0.2.0
langchain-tavily==0.2.0
langgraph==0.2.45
streamlit==1.39.0
python-dotenv==1.0.1
together==1.3.5
tavily-python==0.5.0
httpx==0.27.2